+++ HSTB/kluster/modules/export.py
@@ -7,7 +7,6 @@ from datetime import datetime
 
 from HSTB.kluster.pydro_helpers import is_pydro
 from HSTB.kluster.pdal_entwine import build_entwine_points
-from HSTB.kluster.xarray_helpers import distrib_zarr_write
 
 
 class FqprExport:
